Fixed saving of layout with multiple kpis per chart. Refactored code for TemplatePageElementEditor pass kpi id only without kpi name
